project2/requirements.txt
django
psycopg2-binary
djangorestframework
gunicorn
